Made getTotalCost method for SubRepository and SubsViewModel

+++ app/src/main/java/com/example/subby/SubsDao.java
@@ -24,6 +24,6 @@ public interface SubsDao {
     void deleteSubscription(String name);
 
     @Query("SELECT SUM(price) FROM subs_table")
-    LiveData<Integer> getTotalCost();
+    LiveData<Double> getTotalCost();
 
 }
